requirements.txt
google-cloud-error-reporting==0.30.1
google-cloud-datastore==1.7.3